requirements.txt
Pyrogram
PyYAML
loguru
TgCrypto
SQLAlchemy
rapidocr_onnxruntime